// src/types/filter.inventory.ts
export enum FilterInventory {
  HIGHEST_RARITY = 'Highest Rarity',
  LOWEST_RARITY = 'Lowest Rarity',
  MOST_RECENT = 'Most Recent',
  HIGHEST_REFINEMENT = 'Highest Refinement',
  LOWEST_REFINEMENT = 'Lowest Refinement',
  HIGHEST_DUPLICATES = 'Highest Duplciates',
  LOWEST_DUPLICATES = 'Lowest Duplciates',
  A_TO_Z = 'Sort by (A - Z)',
}